RDCC-5957:Changes made to fix failed test cases
